Works well enough.
